backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.35
python-dotenv==1.0.0
pydantic==2.10.0
pydantic-core==2.27.0
bcrypt==4.0.0
httpx==0.25.0
google-generativeai==0.8.0
